move errorpolicy to importdata; config file tests
